+++ intents/model/entity_test.py
@@ -1,3 +1,5 @@
+import pytest
+
 from intents import Entity
 
 def test_entity_default_name():
@@ -13,3 +15,12 @@ def test_entity_custom_name():
         name = "CustomEntityName"
 
     assert MyEntity.name == 'CustomEntityName'
+
+def test_entity_reserved_name():
+    with pytest.raises(ValueError):
+        class I_MyEntity(Entity):
+            pass
+
+    with pytest.raises(ValueError):
+        class MyEntity(Entity):
+            name = "i_my_entity"